requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-filter==21.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Faker==10.0.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
